refactor: merge into cloudflare.ts
maybe later we split but currently they are categorized per vendor
