+++ src/presets/index.ts
@@ -3,7 +3,6 @@ export * from './azure-functions'
 export * from './azure'
 export * from './base-worker'
 export * from './cloudflare'
-export * from './cloudflare-pages'
 export * from './digital-ocean'
 export * from './firebase'
 export * from './heroku'
